+++ lib/division.dart
@@ -2,7 +2,7 @@ library division;
 
 // parent
 export 'src/widget.dart';
-export 'src/style.dart';
+export 'src/style.dart' show TxtStyle, ParentStyle, Gestures, AngleFormat;
 
 // hex(), rgb() and rgba() function
 export 'src/function/color.dart';